implement ConditionalSearcher
so that we don't have to do hacky things in the scheduler to make sure
that the condition applies.

we ran into this because conditions are only checked before applying
them to the e-graph. because we never apply anything to the e-graph,
we never check rule conditions. by moving the check into the searcher,
we can make sure to always check the conditions.